requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
attrs==20.3.0
awsebcli==3.19.2
base58==2.1.1
bcrypt==3.2.0
blessed==1.18.0
boto3==1.16.51
botocore==1.19.51
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-cors-headers==3.7.0
django-elasticsearch-dsl==7.1.4
django-filter==2.4.0
django-import-export==2.5.0
django-mysql==3.12.0
django-sendgrid-v5==0.9.0
django-storages==1.11.1
djangorestframework==3.12.2
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
et-xmlfile==1.1.0
future==0.16.0
idna==2.10
inflection==0.5.1
ipfshttpclient==0.7.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
MarkupPy==1.14
MarkupSafe==1.1.1
multiaddr==0.0.9
mysqlclient==2.0.3
netaddr==0.8.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.1.0
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-http-client==3.3.1
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
requests-aws4auth==1.0.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.3.7
semantic-version==2.5.0
sendgrid==6.4.8
sendypy==5.2.5
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
tablib==3.0.0
termcolor==1.1.0
texttable==1.6.3
uritemplate==3.0.1
urllib3==1.25.11
varint==1.0.2
videodownloader==2.0.0
wcwidth==0.1.9
websocket-client==0.58.0
xlrd==2.0.1
xlwt==1.3.0
